+++ examples/colors.php
@@ -17,15 +17,15 @@
 
 require_once 'common.php';
 
-\Inane\Cli\Cli::line('  %C%5All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n');
+\Inane\Cli\Cli::line('  %C%5 1. All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n');
 
-echo \Inane\Cli\Colors::colorize('  %C%5All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n', true) . "\n";
-echo \Inane\Cli\Colors::colorize('  %C%5All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n') . "\n";
+echo \Inane\Cli\Colors::colorize('  %C%5 2. All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n', true) . "\n";
+echo \Inane\Cli\Colors::colorize('  %C%5 3. All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n') . "\n";
 
 \Inane\Cli\Colors::enable(); // Forcefully enable
-\Inane\Cli\Cli::line('  %C%5All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n');
+\Inane\Cli\Cli::line('  %C%5 4. All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n');
 
 //\Inane\Cli\Colors::disable(); // Disable forcefully!
-\Inane\Cli\Cli::line('  %C%5All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n', true);
+\Inane\Cli\Cli::line('  %C%5 5. All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n', true);
 \Inane\Cli\Colors::enable(false); // Enable, but not forcefully
-\Inane\Cli\Cli::line('  %C%5All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n');
+\Inane\Cli\Cli::line('  %C%5 6. All output is run through %Y%6\Inane\Cli\Colors::colorize%C%5 before display%n');